6.2 Fixed
Everything works. I provided sample inputs for what the inputs should look like. These use matplotlib again, so i had to use BytesIO and Base64. other than that, everything works. Let me know if you have any questions.
